Add Encryption, Sample User Model
AES encryption in CTR mode is used, using random IVs so that the
encryptor never produces the same cipher text for the same value. The
IV used to encrypt the plaintext is returned prepended to the cipher text.
Erlang’s `crypto:strong_rand_bytes` is used to ensure a cryptographically random
IV is generated.

Because this encryption method doesn’t produce the same value for the
same plaintext, we cannot query encrypted fields through Ecto. If we
try, Ecto will run the search term through the encryptor, which will
produce a new value, and there will be no match in the database.

This can be worked around by adding a second field for those fields we
want to be able to query: `field_name_hash`. This field contains a
sha256 hash of the contents of the `field_name` field. Since sha256
always produces the same value for the same plaintext, we can query on
this field. Therefore, to allow querying, a second `HashField` type is
implemented to transparently hash the search terms when you search on a
given field.

To make this work, it is also necessary to ensure that every time a row
is written to the database, the contents of `field_name` are copied to
`field_name_hash`. See `Encryption.User.set_hashed_fields/1` below, and
the Ecto model callbacks which ensure it is called.

**Security Implications of SHA256 hash fields**

The data cannot be reconstructed from the sha256 hash, though it does
reveal which rows have the same value in that field, since their sha256
hashes will match. Depending on your threat model, this may make this
method of querying the data unusable.